many different changes

+++ MRFHuman/MRFObjCHomework/Source/MRFCarwash/MRFEmployee/MRFEmpoyeeDirector/MRFEmpoyeeDirector.h
@@ -9,8 +9,8 @@
 #import "MRFEmployee.h"
 
 @interface MRFEmpoyeeDirector : MRFEmployee
-@property (nonatomic, assign, readonly)   uint8_t capital;
+@property (nonatomic, assign, readonly)   uint64_t capital;
 
-- (instancetype)initWithCapital:(uint8_t)capital;
+- (instancetype)initWithCapital:(uint64_t)capital;
 
 @end
