+++ MRFHuman/MRFObjCHomework/Source/MRFCarwash/Protocols/MRFMoneyFlow.h
@@ -9,7 +9,7 @@
 #import <Foundation/Foundation.h>
 
 @protocol MRFMoneyFlow <NSObject>
-@property (nonatomic, assign)uint8_t money;
+@property (nonatomic, assign)uint32_t money;
 
 - (void)takeMoney:(uint8_t)money fromMoneyKeeper:(id)moneyKeeper;
